Address thermo-nuclear review follow-ups for PR #26.
Centralize bridge SnapshotApi/ExtensionApi types, add tri-state env parsing, extract abort tracker, collapse bridge suppression into resolveToolCompletion, and fix AGENTS.md module responsibilities.
