+++ src/cursor-pi-tool-bridge-run.ts
@@ -2,14 +2,6 @@ import { randomUUID } from "node:crypto";
 import type { IncomingMessage, ServerResponse } from "node:http";
 import type { McpServerConfig } from "@cursor/sdk";
 import type { Context, ToolResultMessage } from "@earendil-works/pi-ai";
-import type {
-	ExtensionAPI,
-	ExtensionHandler,
-	SessionShutdownEvent,
-	ToolCallEvent,
-	ToolCallEventResult,
-	ToolResultEvent,
-} from "@earendil-works/pi-coding-agent";
 import { Server as McpProtocolServer } from "@modelcontextprotocol/sdk/server/index.js";
 import { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
 import {
@@ -40,6 +32,7 @@ import {
 	snapshotToolToMcpTool,
 	waitForProtocolFlush,
 } from "./cursor-pi-tool-bridge-mcp.js";
+
 export interface CursorPiToolBridgeRunHost {
 	registerRun(pathname: string, run: CursorPiToolBridgeRunImpl): Promise<string>;
 	unregisterRun(pathname: string, run: CursorPiToolBridgeRunImpl): Promise<void>;
@@ -57,140 +50,6 @@ interface PendingBridgeCall {
 	onAbort?: () => void;
 	settled: boolean;
 }
-
-interface CursorPiToolBridgeActiveToolExecution {
-	toolCallId: string;
-	abort: () => Promise<void> | void;
-	cancelPending: (reason: string) => void;
-	signal?: AbortSignal;
-	onAbort?: () => void;
-}
-
-class CursorPiToolBridgeToolExecutionAbortTracker {
-	private readonly activeExecutions = new Map<string, CursorPiToolBridgeActiveToolExecution>();
-	private processSignalHandlersInstalled = false;
-
-	track(
-		toolCallId: string,
-		options: {
-			signal?: AbortSignal;
-			abort: () => Promise<void> | void;
-			cancelPending: (reason: string) => void;
-		},
-	): boolean {
-		this.finish(toolCallId);
-		const execution: CursorPiToolBridgeActiveToolExecution = {
-			toolCallId,
-			abort: options.abort,
-			cancelPending: options.cancelPending,
-			signal: options.signal,
-		};
-		if (options.signal?.aborted) {
-			this.cancelExecution(execution, "Cursor pi bridge tool execution was already aborted");
-			this.abortExecution(execution);
-			return false;
-		}
-
-		execution.onAbort = () => {
-			this.cancelExecution(execution, "Cursor pi bridge tool execution was aborted");
-			this.finish(toolCallId);
-		};
-		execution.signal?.addEventListener("abort", execution.onAbort, { once: true });
-		this.activeExecutions.set(toolCallId, execution);
-		this.installProcessSignalHandlers();
-		return true;
-	}
-
-	finish(toolCallId: string): void {
-		const execution = this.activeExecutions.get(toolCallId);
-		if (!execution) return;
-		if (execution.onAbort) execution.signal?.removeEventListener("abort", execution.onAbort);
-		this.activeExecutions.delete(toolCallId);
-		this.uninstallProcessSignalHandlersIfIdle();
-	}
-
-	finishAll(): void {
-		for (const toolCallId of [...this.activeExecutions.keys()]) this.finish(toolCallId);
-	}
-
-	abortAll(reason: string): void {
-		for (const execution of [...this.activeExecutions.values()]) {
-			this.cancelExecution(execution, reason);
-			this.abortExecution(execution);
-			this.finish(execution.toolCallId);
-		}
-	}
-
-	getActiveCount(): number {
-		return this.activeExecutions.size;
-	}
-
-	emitProcessAbortSignalForTests(signal: NodeJS.Signals): void {
-		this.abortActiveExecutions(signal, { preserveProcessSignalBehavior: true });
-	}
-
-	private readonly handleSigint = (): void => {
-		this.abortActiveExecutions("SIGINT");
-	};
-
-	private readonly handleSigterm = (): void => {
-		this.abortActiveExecutions("SIGTERM");
-	};
-
-	private installProcessSignalHandlers(): void {
-		if (this.processSignalHandlersInstalled) return;
-		this.processSignalHandlersInstalled = true;
-		process.on("SIGINT", this.handleSigint);
-		process.on("SIGTERM", this.handleSigterm);
-	}
-
-	private uninstallProcessSignalHandlersIfIdle(): void {
-		if (!this.processSignalHandlersInstalled || this.activeExecutions.size > 0) return;
-		this.processSignalHandlersInstalled = false;
-		process.off("SIGINT", this.handleSigint);
-		process.off("SIGTERM", this.handleSigterm);
-	}
-
-	private abortActiveExecutions(
-		signal: NodeJS.Signals,
-		options: { preserveProcessSignalBehavior?: boolean } = {},
-	): void {
-		if (this.activeExecutions.size === 0) return;
-		const shouldRestoreDefaultSignalBehavior =
-			options.preserveProcessSignalBehavior !== true && !this.hasExternalProcessSignalListeners(signal);
-		this.abortAll(`Cursor pi bridge tool execution interrupted by ${signal}`);
-		if (shouldRestoreDefaultSignalBehavior) this.restoreDefaultProcessSignalBehavior(signal);
-	}
-
-	private cancelExecution(execution: CursorPiToolBridgeActiveToolExecution, reason: string): void {
-		try {
-			execution.cancelPending(reason);
-		} catch {
-			// Cancellation is best-effort during process abort/shutdown cleanup; keep aborting siblings.
-		}
-	}
-
-	private abortExecution(execution: CursorPiToolBridgeActiveToolExecution): void {
-		try {
-			Promise.resolve(execution.abort()).catch(() => undefined);
-		} catch {
-			// Abort is best-effort during process abort/shutdown cleanup; keep aborting siblings.
-		}
-	}
-
-	private hasExternalProcessSignalListeners(signal: NodeJS.Signals): boolean {
-		const ownHandler = signal === "SIGINT" ? this.handleSigint : this.handleSigterm;
-		return process.listeners(signal).some((listener) => listener !== ownHandler);
-	}
-
-	private restoreDefaultProcessSignalBehavior(signal: NodeJS.Signals): void {
-		setImmediate(() => {
-			process.kill(process.pid, signal);
-		});
-	}
-}
-
-export const bridgeToolExecutionAbortTracker = new CursorPiToolBridgeToolExecutionAbortTracker();
 
 export class CursorPiToolBridgeRunImpl implements CursorPiToolBridgeRun {
 	readonly id: string;
@@ -510,11 +369,3 @@ export class CursorPiToolBridgeRunImpl implements CursorPiToolBridgeRun {
 		if (queuedIndex >= 0) this.queuedRequests.splice(queuedIndex, 1);
 	}
 }
-
-type CursorPiToolBridgeSnapshotApi = Pick<ExtensionAPI, "getActiveTools" | "getAllTools">;
-
-export type CursorPiToolBridgeExtensionApi = CursorPiToolBridgeSnapshotApi & {
-	on(event: "tool_call", handler: ExtensionHandler<ToolCallEvent, ToolCallEventResult>): void;
-	on(event: "tool_result", handler: ExtensionHandler<ToolResultEvent>): void;
-	on(event: "session_shutdown", handler: ExtensionHandler<SessionShutdownEvent>): void;
-};
